Revert "Assume driver version 11.4 on WSL2"
This reverts commit c12ef85213c234fd4c64fc3d1b9ee348f847fa7a.

The WSL2 preview driver (which identifies itself as 510.06/ CUDA 11.6)
only supported PTX up to version 7.4. However, the latest release driver
(496.49) correctly reports the CUDA version it supports (11.5).

Instead of making assumptions about the supported PTX version on WSL
based on incorrect behaviour of the preview driver, we should use the
reported version as normal, and recommend the use of the latest release
driver (and to discontinue use of the preview driver).